requirements.txt
# local package
-e .
jupyter
pandas
autopep8

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
